refactor: change data addaccount implementation name
