+++ src/main/factories/useCases/addAccount/makeDBAddAccount.ts
@@ -1,4 +1,4 @@
-import { DbAddAccount } from '../../../../data/useCases/addaccount/DbAddAccount'
+import { DBAddAccount } from '../../../../data/useCases/addaccount/DBAddAccount'
 import { AddAccount } from '../../../../domain/useCases/AddAccount'
 import { BcrypterAdapter } from '../../../../infra/cryptography/bcrypt/BcryptAdapter'
 import { AccountMongoRepository } from '../../../../infra/db/mongodb/account/AccountMongoRepository'
@@ -7,5 +7,5 @@ export function makeDBAddAccount ():AddAccount {
   const salt = 12
   const encrypter = new BcrypterAdapter(salt)
   const addAccountRepository = new AccountMongoRepository()
-  return new DbAddAccount(encrypter, addAccountRepository, addAccountRepository)
+  return new DBAddAccount(encrypter, addAccountRepository, addAccountRepository)
 }
